transaction-analyzer/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
google-cloud-bigquery==3.16.0
pandas==2.2.1
python-dotenv==1.0.1
